+++ src/Overwurd.Model/Repositories/IReadOnlyOverwurdRepository.cs
@@ -7,7 +7,7 @@ using Overwurd.Model.Models;
 
 namespace Overwurd.Model.Repositories {
     public interface IReadOnlyOverwurdRepository<T> where T : IEntityWithNumericId {
-        Task<T> FindByIdAsync(long id, CancellationToken cancellationToken = default);
+        Task<T> FindByIdAsync(int id, CancellationToken cancellationToken = default);
 
         Task<OverwurdPaginationResult<T>> PaginateAsync(Expression<Func<T, bool>> predicate, int page, int pageSize, CancellationToken cancellationToken = default);
